@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/alecthomas/kingpin"
+	"github.com/alecthomas/kingpin/v2"
 	"github.com/vearutop/plt/curl"
 	"github.com/vearutop/plt/loadgen"
 	"github.com/vearutop/plt/nethttp"
